Former-commit-id: a253fd69885c4177cf116b2c295a16fad53ed924

+++ palladian-retrieval/src/test/java/ws/palladian/retrieval/search/web/VimeoSearcherTest.java
@@ -25,7 +25,7 @@ public class VimeoSearcherTest {
         assertEquals("https://vimeo.com/20800127", result1.getUrl());
         assertEquals(246, (long)result1.getRunTime());
         
-        int resultCount = VimeoSearcher.parseResultCount(jsonString);
+        long resultCount = VimeoSearcher.parseResultCount(jsonString);
         assertEquals(94609, resultCount);
     }
